+++ modules/warp-routing/scripts/generate_subnets.py
@@ -2,7 +2,7 @@ import ipaddress
 import sys
 import json
 import os
-from datetime import datetime
+from datetime import datetime, timezone
 
 def get_script_paths(provider):
     """Return absolute paths for reliable file operations"""
@@ -16,6 +16,7 @@ def get_script_paths(provider):
 
 def validate_input(input_str):
     """Validate and normalize input with comprehensive checks"""
+    # Only add /32 if there's no slash already present
     if '/' not in input_str:
         input_str += '/32'
     
@@ -36,72 +37,60 @@ def validate_input(input_str):
     except ValueError as e:
         raise ValueError(f"Invalid network: {str(e)}")
     
-    return str(network)
+    return network
 
-def infer_base_cidr(exclude_cidr):
-    """Infer base CIDR from RFC1918 ranges with better error reporting"""
-    try:
-        exclude_network = ipaddress.ip_network(exclude_cidr, strict=True)
-    except ValueError as e:
-        raise ValueError(f"Invalid exclusion CIDR: {str(e)}") from e
-    
+def infer_base_cidr(exclude_networks):
+    """Infer common base CIDR from RFC1918 ranges for all exclude networks"""
     rfc1918_ranges = [
         ipaddress.ip_network("10.0.0.0/8"),
         ipaddress.ip_network("172.16.0.0/12"),
         ipaddress.ip_network("192.168.0.0/16"),
     ]
 
-    for base_cidr in rfc1918_ranges:
-        if exclude_network.subnet_of(base_cidr):
-            return str(base_cidr)
-    
-    raise ValueError(f"IP network {exclude_cidr} not in RFC1918 private ranges")
+    # Check all exclude networks are in the same RFC1918 base network
+    base_cidr = None
+    for base in rfc1918_ranges:
+        if all(net.subnet_of(base) for net in exclude_networks):
+            base_cidr = base
+            break
 
-def calculate_exclusions(base_cidr, exclude_cidr):
-    """Enhanced CIDR exclusion with edge case handling"""
-    base = ipaddress.ip_network(base_cidr)
-    exclude = ipaddress.ip_network(exclude_cidr)
+    if base_cidr is None:
+        raise ValueError("All input subnets must be within the same RFC1918 private range")
 
-    # Handle full range exclusion
-    if base == exclude:
-        return []
-    
-    # Handle single IP exclusion
-    if exclude.prefixlen == 32:
-        excluded_ip = int(exclude.network_address)
-        ranges = []
-        
-        if base.network_address < excluded_ip:
-            ranges.append((base.network_address, excluded_ip - 1))
-        if excluded_ip < base.broadcast_address:
-            ranges.append((excluded_ip + 1, base.broadcast_address))
-    else:
-        # Use native exclusion for larger networks
-        try:
-            return [str(net) for net in base.address_exclude(exclude)]
-        except ValueError:
-            return [str(base)]
+    return base_cidr
 
-    # Convert ranges to CIDRs
-    results = []
-    for start, end in ranges:
-        results.extend(ipaddress.summarize_address_range(
-            ipaddress.IPv4Address(start),
-            ipaddress.IPv4Address(end)
-        ))
-    
-    return [str(net) for net in results if net.subnet_of(base)]
+def exclude_subnet_from_list(networks, exclude_net):
+    """Exclude exclude_net from a list of networks, returning new list"""
+    result = []
+    for net in networks:
+        if exclude_net.subnet_of(net):
+            # Exclude from this net
+            excluded_parts = list(net.address_exclude(exclude_net))
+            result.extend(excluded_parts)
+        else:
+            # No overlap, keep as is
+            result.append(net)
+    return result
+
+def calculate_exclusions(base_cidr, exclude_networks):
+    """Exclude multiple subnets from base network"""
+    remaining_networks = [base_cidr]
+
+    for exclude_net in exclude_networks:
+        remaining_networks = exclude_subnet_from_list(remaining_networks, exclude_net)
+
+    return remaining_networks
 
 def main():
-    if len(sys.argv) != 3:
+    if len(sys.argv) < 3:
         error = {
-            "timestamp": datetime.utcnow().isoformat(),
+            "timestamp": datetime.now(timezone.utc).isoformat(),
             "error": "Invalid arguments",
-            "usage": "python3 generate_subnets.py <IP_or_CIDR> <provider>",
-            "example": "python3 generate_subnets.py 10.156.0.23/32 aws",
+            "usage": "python3 script.py <IP_or_CIDR_1> [<IP_or_CIDR_2> ...] <provider>",
+            "example": "python3 script.py 10.156.70.0/24 10.156.82.0/24 gcp",
             "valid_providers": ["aws", "azure", "gcp"]
         }
-        # Attempt to write error to a generic error file if possible
+        # Write to generic error file
         script_dir = os.path.dirname(os.path.abspath(__file__))
         output_dir = os.path.join(script_dir, '..', 'output')
         os.makedirs(output_dir, exist_ok=True)
@@ -110,24 +99,21 @@ def main():
             json.dump(error, f, indent=2)
         sys.exit(1)
 
-    input_str = sys.argv[1].strip()
-    provider = sys.argv[2].strip().lower()
+    *input_strs, provider = sys.argv[1:]
+    provider = provider.lower()
 
     valid_providers = ['aws', 'azure', 'gcp']
     if provider not in valid_providers:
         error = {
-            "timestamp": datetime.utcnow().isoformat(),
-            "input_received": input_str,
+            "timestamp": datetime.now(timezone.utc).isoformat(),
+            "inputs_received": input_strs,
             "error_type": "InvalidProviderError",
             "message": f"Provider '{provider}' is not valid. Must be one of {valid_providers}.",
-            "usage": "python3 generate_subnets.py <IP_or_CIDR> <provider>",
-            "example": "python3 generate_subnets.py 10.156.0.23/32 aws"
+            "usage": "python3 script.py <IP_or_CIDR_1> [<IP_or_CIDR_2> ...] <provider>",
+            "example": "python3 script.py 10.156.70.0/24 10.156.82.0/24 gcp"
         }
-        script_dir = os.path.dirname(os.path.abspath(__file__))
-        output_dir = os.path.join(script_dir, '..', 'output')
-        os.makedirs(output_dir, exist_ok=True)
-        error_path = os.path.join(output_dir, f'{provider}_subnet_error.json')
-        with open(error_path, 'w') as f:
+        paths = get_script_paths(provider)
+        with open(paths['error'], 'w') as f:
             json.dump(error, f, indent=2)
         sys.exit(1)
 
@@ -135,26 +121,25 @@ def main():
     result = {}
 
     try:
-        # Validate and process input
-        exclude_cidr = validate_input(input_str)
-        base_cidr = infer_base_cidr(exclude_cidr)
+        # Validate and process all inputs
+        exclude_networks = [validate_input(s) for s in input_strs]
+        base_cidr = infer_base_cidr(exclude_networks)
         
         # Calculate exclusions
-        excluded_nets = calculate_exclusions(base_cidr, exclude_cidr)
+        excluded_nets = calculate_exclusions(base_cidr, exclude_networks)
         
         # Build result structure
         result = {
             "metadata": {
-                "generated_at": datetime.utcnow().isoformat(),
-                "script_version": "1.1",
-                "input_received": input_str,
-                "normalized_exclusion": exclude_cidr,
-                "base_network": base_cidr,
+                "generated_at": datetime.now(timezone.utc).isoformat(),
+                "script_version": "1.3",
+                "inputs_received": input_strs,
+                "base_network": str(base_cidr),
                 "provider": provider
             },
             "exclusions": [
                 {
-                    "address": cidr,
+                    "address": str(cidr),
                     "description": f"{provider.upper()} Excluded Subnet {i+1}",
                     "type": "calculated"
                 } for i, cidr in enumerate(excluded_nets)
@@ -172,13 +157,13 @@ def main():
 
     except Exception as e:
         error = {
-            "timestamp": datetime.utcnow().isoformat(),
-            "input_received": input_str,
+            "timestamp": datetime.now(timezone.utc).isoformat(),
+            "inputs_received": input_strs,
             "error_type": type(e).__name__,
             "message": str(e),
             "possible_fixes": [
-                "Ensure input is a valid IPv4 address/CIDR",
-                "Check if IP belongs to RFC1918 private ranges",
+                "Ensure all inputs are valid IPv4 addresses/CIDRs",
+                "Check if all IPs belong to the same RFC1918 private range",
                 "Verify network mask validity (0-32)",
                 "Use a valid provider argument: aws, azure, or gcp"
             ]
