--- scripts/aws-browser-ssh-init.tftpl
@@ -1,37 +1,0 @@
-#cloud-config
-hostname: ${hostname}
-package_update: true
-package_upgrade: true
-
-packages:
-  - wget
-  - curl
-  - traceroute
-  - build-essential
-  - hping3
-  - net-tools
-  - unzip
-
-users:
-  - default
-%{ for user in users ~}
-  - name: ${user}
-    passwd: ${user}
-    groups: sudo
-    shell: /bin/bash
-    sudo: ['ALL=(ALL) NOPASSWD:ALL']
-    lock_passwd: false
-%{ endfor ~}
-  - name: ${okta_contractor_username}
-    passwd: ${okta_contractor_password}
-    shell: /bin/bash
-    sudo: ['ALL=(ALL) NOPASSWD:ALL']
-    lock_passwd: false
-
-runcmd:
-  - echo "${ca_cloudflare_browser}" > /etc/ssh/ca_cloudflare_browser.pub
-  - chmod 600 /etc/ssh/ca_cloudflare_browser.pub
-  - sed -i 's/^#PubkeyAuthentication yes/PubkeyAuthentication yes/' /etc/ssh/sshd_config
-  - sed -i '/PubkeyAuthentication yes/a TrustedUserCAKeys /etc/ssh/ca_cloudflare_browser.pub' /etc/ssh/sshd_config
-  - timedatectl set-timezone Europe/Paris
-  - DD_API_KEY=${datadog_api_key} DD_SITE=${datadog_region} bash -c "$(curl -L https://install.datadoghq.com/scripts/install_script_agent7.sh)" > /tmp/dd_install.log 2>&1
