--- scripts/aws-cloudflared-init.yaml
@@ -1,24 +1,0 @@
-#cloud-config
-hostname: ${hostname}
-package_update: true
-package_upgrade: true
-
-packages:
-  - wget
-  - curl
-  - traceroute
-  - build-essential
-  - hping3
-  - net-tools
-  - unzip
-
-runcmd:
-  - sudo mkdir -p --mode=0755 /usr/share/keyrings
-  - curl -fsSL https://pkg.cloudflare.com/cloudflare-main.gpg | sudo tee /usr/share/keyrings/cloudflare-main.gpg >/dev/null
-  - echo 'deb [signed-by=/usr/share/keyrings/cloudflare-main.gpg] https://pkg.cloudflare.com/cloudflared any main' | sudo tee /etc/apt/sources.list.d/cloudflared.list
-  - sudo apt-get update && sudo apt-get install cloudflared
-  - sudo cloudflared service install ${tunnel_secret_aws}
-  - sudo timedatectl set-timezone Europe/Paris
-
-  # Datadog Agent installation
-  - 'DD_API_KEY=${datadog_api_key} DD_SITE=${datadog_region} bash -c "$(curl -L https://install.datadoghq.com/scripts/install_script_agent7.sh)" > /tmp/dd_install.log 2>&1'
