--- scripts/azure-vm-init.tftpl
@@ -1,20 +1,0 @@
-#cloud-config
-hostname: ${hostname}
-package_update: true
-package_upgrade: true
-
-packages:
-  - wget
-  - curl
-  - net-tools
-  - iproute2
-  - nmap
-  - traceroute
-  - build-essential
-  - hping3
-  - unzip
-
-runcmd:
-  - sudo timedatectl set-timezone Europe/Paris
-  # Datadog Agent installation
-  - 'DD_API_KEY=${datadog_api_key} DD_SITE=${datadog_region} bash -c "$(curl -L https://install.datadoghq.com/scripts/install_script_agent7.sh)" > /tmp/dd_install.log 2>&1'
